backend/requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pandas
scikit-learn
mysqlclient
pymysql
python-dotenv
requests
reportlab
cryptography
